Review corrections introduced, CustomTemplate class back to test_base_custom_templates
